+++ ObjectsAsAPI/Models/AtlasRequest.cs
@@ -64,15 +64,17 @@ public partial class AtlasRequest : IRealmObject
         }
     }
 
+    // Used in the UI
     public string? StatusString
     {
         get
         {
+            var response = Response.AsNullableIRealmObject() as IResponse;
             return Status switch
             {
                 RequestStatus.Draft => "Draft",
                 RequestStatus.Pending => "Pending",
-                RequestStatus.Handled => (Response.AsIRealmObject() as IResponse)?.Status.ToString(),
+                RequestStatus.Handled => $"{response?.Status}{(string.IsNullOrEmpty(response?.RejectedReason) ? "" : $": \"{response?.RejectedReason}\"")}",
                 _ => throw new NotImplementedException(),
             };
         }
@@ -132,53 +134,6 @@ public enum ResponseStatus
     Rejected,
 }
 
-public partial class CreateOrderPayload : IRealmObject, IPayload
-{
-    [PrimaryKey]
-    [MapTo("_id")]
-    public ObjectId Id { get; set; } = ObjectId.GenerateNewId();
-
-    [MapTo("_creatorId")]
-    public string CreatorId { get; set; }
-
-    [MapTo("content")]
-    public OrderContent? Content { get; set; }
-
-    public CreateOrderPayload()
-    {
-        if (RealmService.CurrentUser == null)
-        {
-            throw new Exception("Login before using models!");
-        }
-
-        CreatorId = RealmService.CurrentUser.Id;
-    }
-}
-
-public partial class CreateOrderResponse : IRealmObject, IResponse
-{
-    [PrimaryKey]
-    [MapTo("_id")]
-    public ObjectId Id { get; private set; }
-
-    [MapTo("_creatorId")]
-    public string CreatorId { get; private set; } = null!;
-
-    [MapTo("order")]
-    public Order? Order { get; private set; }
-
-    [MapTo("status")]
-    private string _Status { get; set; } = null!;
-
-    public ResponseStatus Status
-    {
-        get => Enum.Parse<ResponseStatus>(_Status);
-        private set => _Status = value.ToString();
-    }
-
-    [MapTo("rejectedReason")]
-    public string? RejectedReason { get; private set; }
-}
 
 /* What I don't like about generic request class:
  *  - Payload/Response can't be embedded, so they need to have an ID and CreatorID
