+++ ObjectsAsAPI/ViewModels/OrderViewModel.cs
@@ -1,6 +1,7 @@
-﻿using System;
-using CommunityToolkit.Mvvm.ComponentModel;
+﻿using CommunityToolkit.Mvvm.ComponentModel;
+using CommunityToolkit.Mvvm.Input;
 using ObjectsAsAPI.Models;
+using ObjectsAsAPI.Services;
 
 namespace ObjectsAsAPI.ViewModels;
 
@@ -9,5 +10,29 @@ public partial class OrderViewModel : BaseViewModel
 {
     [ObservableProperty]
     private Order _order = null!;
+
+    [RelayCommand]
+    public async Task CancelOrder()
+    {
+        var realm = RealmService.GetMainThreadRealm();
+
+        var requestPayload = new CancelOrderPayload
+        {
+            OrderId = Order.Id,
+        };
+
+        var request = new AtlasRequest
+        {
+            Status = RequestStatus.Pending,
+            Payload = requestPayload,
+        };
+
+        realm.Write(() =>
+        {
+            realm.Add(request);
+        });
+
+        await Shell.Current.GoToAsync("..");
+    }
 }
